requirements.txt
Django>=4.1,<5.0            
psycopg2-binary>=2.9         

django-environ>=0.9.0       

gunicorn>=20.1.0            
whitenoise>=6.0.0           

django-extensions>=3.2.0    
django-cors-headers>=3.13.0 
django-debug-toolbar>=3.5.0 
djangorestframework>=3.14.0 
django-crispy-forms>=1.14.0 
pytest>=7.2.0               
pytest-django>=4.5.2       
flake8>=6.0.0               
